requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argcomplete==3.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.19.0
azureml-mlflow==1.55.0
Babel==2.15.0
beautifulsoup4==4.12.3
bitsandbytes==0.41.1
black==24.4.2
bleach==6.1.0
blinker==1.7.0
bnlearn==0.8.7
cachetools==5.3.1
causalgen==0.1.1
causica==0.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.1
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
Cython==0.29.34
dataclasses-json==0.5.14
datazets==0.1.9
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
df2onehot==1.0.7
distro==1.7.0
distro-info===1.1build1
docker==7.0.0
docstring_parser==0.16
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
funcsigs==1.0.2
gast==0.4.0
gbinder-python==1.1.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.54.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.8.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hydra-core==1.3.2
idna==3.4
immutabledict==4.2.0
importlib-metadata==4.6.4
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
ismember==1.0.3
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.11
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonargparse==4.20.1
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kaggle==1.6.8
keras==2.12.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
lightning-utilities==0.11.2
lit==18.1.2
Mako==1.3.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mlflow==2.11.3
mlflow-skinny==2.11.3
more-itertools==8.10.0
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.3.0
opt-einsum==3.3.0
ortools==9.9.3963
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pgmpy==0.1.25
pillow==10.2.0
platformdirs==4.2.1
portalocker==2.8.2
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.12.0
pybnesian==0.4.3
pycparser==2.22
pydot==1.4.2
pyelftools==0.30
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pypickle==1.1.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-louvain==0.16
python-slugify==8.0.4
pytorch-lightning==2.1.2
pytz==2024.1
PyYAML==5.4.1
pyzed @ file:///tmp/selfgz683/pyzed-4.0-cp310-cp310-linux_x86_64.whl
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
querystring-parser==1.2.4
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.10.1
SCons==4.0.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
systemd-python==234
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboardX==2.6.2.2
tensordict==0.1.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.4.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
torch==2.3.0
torchaudio==2.2.1
torchmetrics==1.3.2
torchsummary==1.5.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
triton==2.3.0
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240311
typeshed_client==2.5.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.16
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.14.1
xkit==0.0.0
xlrd==2.0.1
yarl==1.9.4
zipp==1.0.0
